# public/openapi.yaml
openapi: 3.0.3
info:
  title: E-Puskesmas EIS API
  version: 1.0.0
  description: |
    OpenAPI untuk E-Puskesmas EIS. Disusun mengikuti struktur menu frontend:
    Dashboard, Cari, Klaster 1-4, Lintas Klaster, Monitoring, dan Akun.

    **Integrasi pihak ketiga (e-Puskesmas / webhook)**
    - Semua pengiriman data dari sistem eksternal dilakukan via endpoint `POST /api/eis/klaster1`, `POST /api/eis/klaster2`, `POST /api/eis/klaster3`, `POST /api/eis/klaster4`, dan `POST /api/eis/lintas-klaster`.
    - Endpoint dashboard/analitik tetap `GET` (read-only). Data yang tampil berasal dari hasil POST (webhook push).
    - Semua endpoint POST di atas wajib menyertakan header `X-Epus-Signature` untuk HMAC:
      signature = HMAC_SHA256(EPUS_WEBHOOK_SECRET, `${t}.${rawJsonUtf8}`)
      Header format: `t=<unix_ts>,v1=<hex_sha256>`
    - Idempotency: gunakan `batchId` unik agar retry aman (tidak dobel).
    - Skala besar: payload boleh `Content-Encoding: gzip` (signature dihitung dari JSON *setelah* di-decompress).
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  parameters:
    XEpusSignature:
      name: X-Epus-Signature
      in: header
      required: true
      description: "Signature HMAC. Format: `t=<unix_ts>,v1=<hex_sha256>`"
      schema:
        type: string
        example: t=1700000000,v1=0123abcd...
    ContentEncoding:
      name: Content-Encoding
      in: header
      required: false
      description: "Opsional untuk payload besar. Gunakan `gzip` bila body di-compress."
      schema:
        type: string
        enum: [gzip]
...
    PuskesmasId:
      name: puskesmasId
      in: query
      required: false
      schema:
        type: string
      description: "ID puskesmas, gunakan `all` untuk semua"
    Bulan:
      name: bulan
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 12
      description: Bulan (1-12)
    Tahun:
      name: tahun
      in: query
      required: false
      schema:
        type: integer
      description: Tahun (YYYY)
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
...
    KlasterGenericResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          additionalProperties: true
...
paths:
...
  /api/eis/klaster3:
    get:
      tags: [Klaster 3]
      summary: Data Klaster 3
      parameters:
        - $ref: '#/components/parameters/PuskesmasId'
        - $ref: '#/components/parameters/Bulan'
        - $ref: '#/components/parameters/Tahun'
      responses:
        '200':
          description: Data klaster 3
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KlasterGenericResponse'
    post:
      tags: [Klaster 3]
      summary: Webhook push data Klaster 3 (Deteksi Dini, Faktor Risiko, Pemeriksaan Gigi)
      description: |
        Contoh kasus: **Puskesmas Bumiayu** ingin mengirim data Klaster 3 untuk **Pemeriksaan Gigi**.

        Pihak e-Puskesmas mengirim agregat bulanan per `kategori` (mis. `SD/MI`, `SMP/MTs`, `SMA/SMK`).
        Server akan **upsert** berdasarkan kombinasi unik:
        - Deteksi Dini: `puskesmas + jenis + bulan + tahun`
        - Faktor Risiko: `puskesmas + faktor + kelompokUmur + bulan + tahun`
        - Pemeriksaan Gigi: `puskesmas + kategori + bulan + tahun`
      parameters:
        - $ref: '#/components/parameters/XEpusSignature'
        - $ref: '#/components/parameters/ContentEncoding'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                batchId: { type: string }
                puskesmasKode: { type: string }
                puskesmasId: { type: string }
                bulan: { type: integer, minimum: 1, maximum: 12 }
                tahun: { type: integer }
                deteksiDini:
                  type: array
                  items:
                    type: object
                    required: [jenis, sasaran, capaian]
                    properties:
                      jenis: { type: string }
                      sasaran: { type: integer }
                      capaian: { type: integer }
                faktorRisiko:
                  type: array
                  items:
                    type: object
                    required: [faktor, kelompokUmur, jumlahKasus]
                    properties:
                      faktor: { type: string }
                      kelompokUmur: { type: string }
                      jumlahKasus: { type: integer }
                pemeriksaanGigi:
                  type: array
                  items:
                    type: object
                    required: [kategori, sasaran, diperiksa, perluPerawatan]
                    properties:
                      kategori: { type: string }
                      sasaran: { type: integer }
                      diperiksa: { type: integer }
                      perluPerawatan: { type: integer }
              oneOf:
                - required: [batchId, puskesmasKode, bulan, tahun]
                - required: [batchId, puskesmasId, bulan, tahun]
            examples:
              full_deteksi_faktor_gigi:
                summary: Kirim Deteksi Dini + Faktor Risiko + Pemeriksaan Gigi
                value:
                  batchId: EPUS-BUMIAYU-2026-02-K3-0001
                  puskesmasKode: "3329XXXX"
                  bulan: 2
                  tahun: 2026
                  deteksiDini:
                    - jenis: Kanker Serviks (IVA)
                      sasaran: 200
                      capaian: 120
                    - jenis: Tuberculosis
                      sasaran: 150
                      capaian: 80
                  faktorRisiko:
                    - faktor: Hipertensi
                      kelompokUmur: 45-59
                      jumlahKasus: 42
                    - faktor: Diabetes
                      kelompokUmur: 60+
                      jumlahKasus: 18
                  pemeriksaanGigi:
                    - kategori: SD/MI
                      sasaran: 120
                      diperiksa: 95
                      perluPerawatan: 22
                    - kategori: SMP/MTs
                      sasaran: 80
                      diperiksa: 60
                      perluPerawatan: 15
              only_pemeriksaan_gigi:
                summary: Kirim Pemeriksaan Gigi saja
                value:
                  batchId: EPUS-BUMIAYU-2026-02-K3-GIGI-0001
                  puskesmasKode: "3329XXXX"
                  bulan: 2
                  tahun: 2026
                  pemeriksaanGigi:
                    - kategori: SD/MI
                      sasaran: 120
                      diperiksa: 95
                      perluPerawatan: 22
      responses:
        '200':
          description: Data diterima dan diproses
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '400':
          description: Validasi gagal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Signature invalid
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Kesalahan server
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...